chore: rename files and move promise examples to legacy folders
- Renamed various JavaScript example files for clarity
- Moved general promise examples to `javascript/legacy/js-code-snippets`
- Moved `check-inventory-promises` project to `javascript/legacy/`
