+++ javascript/legacy/check-inventory-promises/README.md
