+++ javascript/legacy/js-code-snippets/promises/complete-a-promise-with-resolve-and-reject.js
