+++ javascript/legacy/js-code-snippets/promises/create-a-javascript-promise.js
