+++ javascript/legacy/js-code-snippets/promises/handle-a-fulfilled-promise-with-then.js
