+++ javascript/legacy/js-code-snippets/promises/handle-a-rejected-promise-with-catch.js
